supply RSTUDIO_PANDOC to packaging tasks, including R Markdown render
